requirements.txt
langchain
langchain-community
groq
langchain-groq
python-dotenv
streamlit